requirements.txt
tqdm==4.32.2
pycrypto==2.6.1
requests==2.22.0
beautifulsoup4==4.7.1
